requirements.txt
fastapi
uvicorn
requests
python-dotenv
streamlit
openai